+++ compose.yaml
@@ -17,6 +17,7 @@ services:
     ##ENV GOOGLE_APPLICATION_CREDENTIALS '~/.config/gcloud/application_default_credentials.json'
     environment:
       - 'GOOGLE_APPLICATION_CREDENTIALS=/app/.config/gcloud/application_default_credentials.json'
+      - 'CLOUDPROXY=cloudsql-proxy:5432'
     volumes:
       - type: bind
         source: ${HOME}/.config/gcloud
@@ -28,11 +29,11 @@ services:
       - '8080:8080'
     depends_on:
       - cloudsql-proxy
-      - flyway
-  flyway:
-    image: flyway/flyway
-    command: -url=jdbc:postgresql://cloudsql-proxy:5432/eventdb -schemas=eventdb -user=postgres -password=secret -connectRetries=60 migrate
-    volumes:
-      - ./src/main/resources/db/migration/:/flyway/sql/
-    depends_on:
-      - cloudsql-proxy
+#      - flyway
+#  flyway:
+#    image: flyway/flyway
+#    command: -url=jdbc:postgresql://cloudsql-proxy:5432/eventdb -schemas=eventdb -user=postgres -password=secret -connectRetries=60 migrate
+#    volumes:
+#      - ./src/main/resources/db/migration/:/flyway/sql/
+#    depends_on:
+#      - cloudsql-proxy
