+++ pom.xml
@@ -51,7 +51,11 @@
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
         </dependency>
-
+        <dependency>
+            <groupId>com.google.cloud.sql</groupId>
+            <artifactId>postgres-socket-factory</artifactId>
+            <version>1.19.0</version>
+        </dependency>
     </dependencies>
 
     <dependencyManagement>
@@ -108,6 +112,7 @@
                 <artifactId>flyway-maven-plugin</artifactId>
                 <version>10.7.1</version>
                 <configuration>
+
                     <user>postgres</user>
                     <password>secret</password>
                     <schemas>
