+++ src/main/resources/application.properties
@@ -3,25 +3,43 @@ spring.cloud.gcp.project-id=tr4-2024-06-deutschebank
 
 server.port=${PORT:8080}
 
+check-tickets-application.cloud-proxy=${CLOUDPROXY:34.147.221.142:5432}
+check-tickets-application.cloud-thing=jdbc:postgresql://${check-tickets-application.cloud-proxy}/eventdb?cloudSqlInstance=${spring.cloud.gcp.sql.instance-connection-name}&socketFactory=com.google.cloud.sql.postgres.SocketFactory
+#  &enableIamAuth=true&user=postgres&password=secret&user=p327570643859-s9yq97@gcp-sa-cloud-sql.iam
+##cloudsql-proxy:5432
+
 #todo GOOGLE_CLOUD_PROJECT env variable
 #projects/tr4-2024-06-deutschebank/topics/
 #todo GOOGLE_APPLICATION_CREDENTIALS environment variable.
 
+spring.cloud.gcp.sql.instance-connection-name=tr4-2024-06-deutschebank:europe-west2:myinstance
+spring.cloud.gcp.sql.database-name=eventdb
+spring.cloud.gcp.sql.jdbc.enabled=true
 
 spring.application.name=check-tickets-application
-spring.datasource.url=jdbc:postgresql://cloudsql-proxy:5432/eventdb
+#spring.datasource.url=jdbc:postgresql://cloudsql-proxy:5432/eventdb
 spring.datasource.username=postgres
 spring.datasource.password=secret
 spring.flyway.user=postgres
 spring.flyway.password=secret
 spring.flyway.schemas=eventsdb
 spring.flyway.locations=filesystem:src/main/resources/db/migration
-spring.flyway.url=jdbc:postgresql://cloudsql-proxy:5432/eventdb
-jakarta.persistence.jdbc.url=jdbc:postgresql://cloudsql-proxy:5432/eventdb
+spring.flyway.connect-retries=5
+#spring.flyway.url=jdbc:postgresql://cloudsql-proxy:5432/eventdb
+#jakarta.persistence.jdbc.url=jdbc:postgresql://cloudsql-proxy:5432/eventdb
+spring.flyway.url=${check-tickets-application.cloud-thing}
+#spring.flyway.url=jdbc:postgresql://5432:5432/eventdb
+jakarta.persistence.jdbc.url=jdbc:postgresql://${check-tickets-application.cloud-proxy}/eventdb
 
 spring.jpa.hibernate.ddl.auto=none
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.show-sql=true
+#327570643859-compute@developer.gserviceaccount.com
+spring.datasource.url=${check-tickets-application.cloud-thing}
+##spring.datasource.url=jdbc:postgresql://google/mydb?cloudSqlInstance=myInstance&socketFactory=com.google.cloud.sql.postgres.SocketFactory&enableIamAuth=true
+
+spring.datasource.hikari.connection-timeout=20000
+
 
 #springdoc.api-docs.path=/api-docs
 #springdoc.packages-to-scan=com.example.todo_app.controller
